requirements.txt
django==6.0.2
djangorestframework==3.16.1
openpyxl==3.1.5
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.6
